Struct google_appengine1::AutomaticScaling

pub struct AutomaticScaling {
    pub cpu_utilization: Option<CpuUtilization>,
    pub network_utilization: Option<NetworkUtilization>,
    pub min_idle_instances: Option<i32>,
    pub max_pending_latency: Option<String>,
    pub max_idle_instances: Option<i32>,
    pub disk_utilization: Option<DiskUtilization>,
    pub request_utilization: Option<RequestUtilization>,
    pub cool_down_period: Option<String>,
    pub max_total_instances: Option<i32>,
    pub max_concurrent_requests: Option<i32>,
    pub min_total_instances: Option<i32>,
    pub min_pending_latency: Option<String>,
}

Automatic scaling is based on request rate, response latencies, and other application metrics.

This type is not used in any activity, and only used as part of another schema.

Fields
cpu_utilization: Option<CpuUtilization>

Target scaling by CPU usage.

network_utilization: Option<NetworkUtilization>

Target scaling by network usage.

min_idle_instances: Option<i32>

Minimum number of idle instances that should be maintained for this version. Only applicable for the default version of a service.

max_pending_latency: Option<String>

Maximum amount of time that a request should wait in the pending queue before starting a new instance to handle it.

max_idle_instances: Option<i32>

Maximum number of idle instances that should be maintained for this version.

disk_utilization: Option<DiskUtilization>

Target scaling by disk usage.

request_utilization: Option<RequestUtilization>

Target scaling by request utilization.

cool_down_period: Option<String>

Amount of time that the Autoscaler should wait between changes to the number of virtual machines. Only applicable for VM runtimes.

max_total_instances: Option<i32>

Maximum number of instances that should be started to handle requests.

max_concurrent_requests: Option<i32>

Number of concurrent requests an automatic scaling instance can accept before the scheduler spawns a new instance. Defaults to a runtime-specific value.

min_total_instances: Option<i32>

Minimum number of instances that should be maintained for this version.

min_pending_latency: Option<String>

Minimum amount of time a request should wait in the pending queue before starting a new instance to handle it.