Struct google_appengine1::Instance

pub struct Instance {
    pub errors: Option<i32>,
    pub name: Option<String>,
    pub app_engine_release: Option<String>,
    pub vm_id: Option<String>,
    pub memory_usage: Option<String>,
    pub availability: Option<String>,
    pub average_latency: Option<i32>,
    pub vm_debug_enabled: Option<bool>,
    pub start_time: Option<String>,
    pub qps: Option<f32>,
    pub requests: Option<i32>,
    pub vm_name: Option<String>,
    pub id: Option<String>,
    pub vm_zone_name: Option<String>,
    pub vm_status: Option<String>,
}

An Instance resource is the computing unit that App Engine uses to automatically scale an application.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

services versions instances get apps (response)

Fields
errors: Option<i32>

Number of errors since this instance was started. @OutputOnly

name: Option<String>

Full path to the Instance resource in the API. Example: apps/myapp/services/default/versions/v1/instances/instance-1. @OutputOnly

app_engine_release: Option<String>

App Engine release this instance is running on. @OutputOnly

vm_id: Option<String>

Virtual machine ID of this instance. Only applicable for instances in App Engine flexible environment. @OutputOnly

memory_usage: Option<String>

Total memory in use (bytes). @OutputOnly

availability: Option<String>

Availability of the instance. @OutputOnly

average_latency: Option<i32>

Average latency (ms) over the last minute. @OutputOnly

vm_debug_enabled: Option<bool>

Whether this instance is in debug mode. Only applicable for instances in App Engine flexible environment. @OutputOnly

start_time: Option<String>

Time that this instance was started. @OutputOnly

qps: Option<f32>

Average queries per second (QPS) over the last minute. @OutputOnly

requests: Option<i32>

Number of requests since this instance was started. @OutputOnly

vm_name: Option<String>

Name of the virtual machine where this instance lives. Only applicable for instances in App Engine flexible environment. @OutputOnly

id: Option<String>

Relative name of the instance within the version. Example: instance-1. @OutputOnly

vm_zone_name: Option<String>

Zone where the virtual machine is located. Only applicable for instances in App Engine flexible environment. @OutputOnly

vm_status: Option<String>

Status of the virtual machine where this instance lives. Only applicable for instances in App Engine flexible environment. @OutputOnly